+++ driver/web/docs/gen/gen_types.go
@@ -373,6 +373,9 @@ type UserProfile struct {
 	PhotoUrl  *string `json:"photo_url,omitempty"`
 }
 
+// PostAdminApplicationJSONBody defines parameters for PostAdminApplication.
+type PostAdminApplicationJSONBody Application
+
 // PostAdminGlobalConfigJSONBody defines parameters for PostAdminGlobalConfig.
 type PostAdminGlobalConfigJSONBody GlobalConfig
 
@@ -434,6 +437,9 @@ type GetTpsServiceRegsParams struct {
 	// A comma-separated list of service IDs to return registrations for
 	Ids string `json:"ids"`
 }
+
+// PostAdminApplicationJSONRequestBody defines body for PostAdminApplication for application/json ContentType.
+type PostAdminApplicationJSONRequestBody PostAdminApplicationJSONBody
 
 // PostAdminGlobalConfigJSONRequestBody defines body for PostAdminGlobalConfig for application/json ContentType.
 type PostAdminGlobalConfigJSONRequestBody PostAdminGlobalConfigJSONBody