+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Set up OIDC compliant token validation endpoints [#51](https://github.com/rokwire/core-building-block/issues/51)
+- Storage improvements [#144](https://github.com/rokwire/core-building-block/issues/144)
+- Expose admin API which creates application [#82](https://github.com/rokwire/core-building-block/issues/82)
 - Suppress logs from the AWS load balancer health checks [#141](https://github.com/rokwire/core-building-block/issues/141)
 - Set up accounts [#18](https://github.com/rokwire/core-building-block/issues/18)
 - Optional OIDC URL overrides [#139](https://github.com/rokwire/core-building-block/issues/139)